+++ com.templlo.service.review/src/main/java/com/templlo/service/review/external/feignClient/client/ReservationClient.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.templlo.service.review.common.response.ApiResponse;
-import com.templlo.service.review.external.feignClient.FeignConfiguration;
+import com.templlo.service.review.external.feignClient.config.FeignConfiguration;
 import com.templlo.service.review.external.feignClient.dto.ReservationData;
 
 @FeignClient(
